first attempt to fix styles not being created/enqueued on the front-end
