+++ src/blocks/member-info/block.json
@@ -23,8 +23,9 @@
             "default": true
         }
     },
-    "render": "file:./render.php",
-    "editorScript": "memberlite-block-member-info-editor",
-    "editorStyle": "file:./index.css",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css",
+	"render": "file:./render.php",
     "textdomain": "memberlite"
 }
