+++ src/blocks/nav-menu/block.json
@@ -39,8 +39,9 @@
             "default": "primary"
         }
     },
-    "render": "file:./render.php",
-    "editorScript": "memberlite-block-nav-menu-editor",
-    "editorStyle": "file:./index.css",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css",
+	"render": "file:./render.php",
     "textdomain": "memberlite"
 }
